requirements.txt
numpy==1.20.2
Pillow==8.2.0
pyquaternion==0.9.9
